Modified Task; missing some getters; name also wasn't being initialized in the constructor.
Fixed the queries in TaskTest.

Added a separate class that test that the user and task created in the
test classes have been deleted. JUnit doesn't always run the tests in
the same order and therefore would sometimes yeild a null pointer
exception. It would delete the user before updating it.

The database has also slightly changed Project_id and task_id are not
auto increment anymore.